+++ WeatherForecast/WeatherForecast/Data/DTO/CloudsDTO.swift
@@ -1,0 +1,12 @@
+//
+//  CloudsDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct CloudsDTO: Decodable {
+    let all: Int?
+}
