+++ WeatherForecast/WeatherForecast/Data/DTO/CoordinateDTO.swift
@@ -1,0 +1,13 @@
+//
+//  CoordinateDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct CoordinateDTO: Decodable {
+    let lon: Double?
+    let lat: Double?
+}
