+++ WeatherForecast/WeatherForecast/Data/DTO/ForecastRainDTO.swift
@@ -1,0 +1,16 @@
+//
+//  ForecastRainDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct ForecastRainDTO: Decodable {
+    let the3H: Double
+    
+    enum CodingKeys: String, CodingKey {
+        case the3H = "3h"
+    }
+}
