+++ WeatherForecast/WeatherForecast/Data/DTO/ForecastSystemDTO.swift
@@ -1,0 +1,12 @@
+//
+//  ForecastSystemDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct ForecastSystemDTO: Decodable {
+    let pod: String
+}
