+++ WeatherForecast/WeatherForecast/Data/DTO/TemperatureInformationDTO.swift
@@ -1,0 +1,23 @@
+//
+//  TemperatureInformationDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct TemperatureInformationDTO: Decodable {
+    let temp, feelsLike, tempMin, tempMax: Double?
+    let pressure, humidity, seaLevel, grndLevel: Int?
+    
+    enum CodingKeys: String, CodingKey {
+        case temp
+        case feelsLike = "feels_like"
+        case tempMin = "temp_min"
+        case tempMax = "temp_max"
+        case pressure, humidity
+        case seaLevel = "sea_level"
+        case grndLevel = "grnd_level"
+    }
+}
