+++ WeatherForecast/WeatherForecast/Data/DTO/WindDTO.swift
@@ -1,0 +1,13 @@
+//
+//  WindDTO.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct WindDTO: Decodable {
+    let speed, gust: Double?
+    let deg: Int?
+}
