+++ WeatherForecast/WeatherForecast/Domain/Entities/CityEntity.swift
@@ -1,0 +1,12 @@
+//
+//  CityEntity.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct CityEntity {
+    let name, country: String
+}
