+++ WeatherForecast/WeatherForecast/Domain/Entities/ForecastEntity.swift
@@ -8,25 +8,6 @@
 import Foundation
 
 struct ForecastEntity {
-    let list: [ListEntity]
+    let list: [ForecastListEntity]
     let city: CityEntity
-}
-
-struct CityEntity {
-    let name, country: String
-}
-
-struct ListEntity {
-    let dtTxt: String
-    let main: MainForecastEntity
-    let weather: [FiveDaysWeatherElementEntity]
-}
-
-struct MainForecastEntity {
-    let temp: Double?
-}
-
-struct FiveDaysWeatherElementEntity {
-    let id: Int?
-    let main, description, icon: String?
 }
