+++ WeatherForecast/WeatherForecast/Domain/Entities/TempEntity.swift
@@ -1,0 +1,12 @@
+//
+//  TempEntity.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct TempEntity {
+    let temp: Double?
+}
