+++ WeatherForecast/WeatherForecast/Domain/Entities/WeatherInformationEntity.swift
@@ -1,0 +1,13 @@
+//
+//  WeatherInformationEntity.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/04.
+//
+
+import Foundation
+
+struct WeatherInformationEntity {
+    let id: Int?
+    let main, description, icon: String?
+}
